Split client and common source objects
